+++ examples/route_guide/client.rs
@@ -21,8 +21,16 @@ extern crate rand;
 #[macro_use]
 extern crate serde_derive;
 extern crate serde_json;
+#[macro_use(slog_o, slog_kv)]
+extern crate slog;
+extern crate slog_async;
+extern crate slog_stdlog;
+extern crate slog_scope;
+extern crate slog_term;
 
 mod util;
+#[path = "../log_util.rs"]
+mod log_util;
 
 use std::sync::Arc;
 use std::time::Duration;
@@ -161,7 +169,7 @@ fn route_chat(client: &RouteGuideClient) {
 }
 
 fn main() {
-    let _guard = grpcio_proto::util::init_log();
+    let _guard = log_util::init_log();
     let env = Arc::new(Environment::new(2));
     let channel = ChannelBuilder::new(env).connect("127.0.0.1:50051");
     let client = RouteGuideClient::new(channel);
