+++ proto/Cargo.toml
@@ -1,20 +1,21 @@
 [package]
 name = "grpcio-proto"
-version = "0.1.0"
+version = "0.1.2"
+authors = ["The TiKV Project Developers"]
+license = "Apache-2.0"
+keywords = ["grpc", "proto"]
+repository = "https://github.com/pingcap/grpc-rs"
+homepage = "https://github.com/pingcap/grpc-rs"
+documentation = "https://docs.rs/grpcio-sys"
+description = "Public proto files for grpcio."
+categories = ["network-programming"]
 build = "build.rs"
-publish = false
 
 [dependencies]
 protobuf = "1.2"
 futures = "0.1"
-grpcio = { path = ".." }
-log = "0.3"
-slog = "2.0"
-slog-async = "2.1"
-slog-stdlog = "3.0"
-slog-scope = "4.0"
-slog-term = "2.2"
+grpcio = { path = "..", version = "0.1.2" }
 
 [build-dependencies]
 protobuf = "1.2"
-grpcio-compiler = { path = "../compiler" }
+grpcio-compiler = { path = "../compiler", version = "0.1.0" }
